Remove Rc from Config and use clones instead
This is another attempt at making our config changeable, but I'm not really happy at how its panning out.

This kind of works but I've realised that the coordinator also needs to know when the config changes, so that inverters_for_message operates correctly.

Might be time for a more fundamental architecture rethink.